addlogo-methods {pixmap}	R Documentation

Methods for Adding a Pixmap Logo to a Plot

Description

This method allows the addition of a re-scaled pixmap to an existing plot, possibly as a logo, optionally preserving aspect. It may be used interactively with locator, and is positioned in the coordinate system of the plot region. Since the logo is displayed using image, it does not yet seem possible to use the function outside this region.

Usage

  addlogo(x, ...)
  ## S4 method for signature 'pixmap'
addlogo(x, px, py, asp = NULL)

Arguments

x
an object of class pixmap

px
a vector of two x coordinates, or a list with two named elements x and y, such as that returned by locator.

py
if px is not a list, a vector of two y coordinates

asp
if omitted or NULL (default), output respects both y coordinates, if a number greater than zero, aspect is preserved in proportion to the difference between x coordinates, multiplied by asp, and only the first y coordinate is respected.

...
potentially further arguments passed to and from methods.

Value

The same pixmap object with changed bounding box and cell resolution

Author(s)

Roger Bivand

Examples

x <- read.pnm(system.file("pictures/logo.ppm", package="pixmap")[1])
plot(x)
for (i in 1:7)
  addlogo(x, px=c(0, (101/77)*11), py=c((i-1)*11, i*11), asp=1)

[Package pixmap version 0.4-13 Index]