channels-methods {pixmap}	R Documentation

Methods for Channel Manipulation

Description

Functions for manipulation and extraction of colors from channel-based pixmap formats. Methods for generic addChannels adds the color channels of a colored pixmap and returns a grey version. Methods for generic getChannels return numeric matrices or arrays containing the specified channels.

Usage

  addChannels(object, coef = NULL)
  getChannels(object, colors = "all")

Arguments

object
Object of class "pixmap".

coef
Coefficients for the color channels, a numeric vercot with as many elements as there are color channels in the pixmap. The default for RGB is c(0.30, 0.59, 0.11), which makes a luminance-based conversion from color to grey.

colors
Character vector naming the color channels that shall be extracted. The default of "all" returns all channels simultaneously.

Author(s)

Friedrich Leisch

Examples

  x <- pixmapRGB(rep(1:5, 3), nrow=4)
  plot(x)
  print(x)

  getChannels(x)
  getChannels(x, colors=c("red", "green"))

  y = addChannels(x)
  plot(y)
  print(y)

  ## extract only the red channel
  y = addChannels(x, coef=c(1,0,0))
  plot(y)

[Package pixmap version 0.4-13 Index]